Create NeuralNetwork.py
Further testing needs to be done, but it works with both the flowers, and dosage networks, granted they are both small, but they are basically the only ones that I can check if they are working right now until I find more. This could be added to by having more types of activation functions like sigmoid, ect. 

I also want to try to speed up the effenciy later down the line